creating http_utils; moving keys_utils and testing
